requirements.txt
pylint==1.9.3
apache-airflow==1.10.0
pylint-quotes==0.1.8
